+++ src/main/java/net/silentchaos512/gear/client/util/GearClientHelper.java
@@ -120,6 +120,11 @@ public final class GearClientHelper {
             TextListBuilder builder = new TextListBuilder();
 
             for (ItemStat stat : displayStats) {
+                if (stat == ItemStats.ENCHANTABILITY && !Config.Common.allowEnchanting.get()) {
+                    // Enchanting not allowed, so hide the stat
+                    continue;
+                }
+
                 float statValue = GearData.getStat(stack, stat);
 
                 StatInstance inst = StatInstance.of(statValue, StatInstance.Operation.AVG, StatInstance.DEFAULT_KEY);