+++ src/main/java/net/silentchaos512/gear/item/gear/CoreArmor.java
@@ -179,7 +179,7 @@ public class CoreArmor extends DyeableArmorItem implements ICoreArmor {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override