+++ src/main/java/net/silentchaos512/gear/item/gear/CoreArrow.java
@@ -190,7 +190,7 @@ public class CoreArrow extends ArrowItem implements ICoreItem {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override