+++ src/main/java/net/silentchaos512/gear/item/gear/CoreBow.java
@@ -184,7 +184,7 @@ public class CoreBow extends BowItem implements ICoreRangedWeapon {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override