+++ src/main/java/net/silentchaos512/gear/item/gear/CoreCrossbow.java
@@ -388,7 +388,7 @@ public class CoreCrossbow extends CrossbowItem implements ICoreRangedWeapon {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override