+++ src/main/java/net/silentchaos512/gear/item/gear/CoreCurio.java
@@ -122,7 +122,7 @@ public class CoreCurio extends Item implements ICoreItem {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override