+++ src/main/java/net/silentchaos512/gear/item/gear/CorePickaxe.java
@@ -131,7 +131,7 @@ public class CorePickaxe extends PickaxeItem implements ICoreTool {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override