+++ src/main/java/net/silentchaos512/gear/item/gear/CoreShears.java
@@ -103,7 +103,7 @@ public class CoreShears extends ShearsItem implements ICoreTool {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override