+++ src/main/java/net/silentchaos512/gear/item/gear/CoreShield.java
@@ -118,7 +118,7 @@ public class CoreShield extends ShieldItem implements ICoreItem {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override