+++ src/main/java/net/silentchaos512/gear/item/gear/CoreShovel.java
@@ -132,7 +132,7 @@ public class CoreShovel extends ShovelItem implements ICoreTool {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override