+++ src/main/java/net/silentchaos512/gear/item/gear/CoreSickle.java
@@ -278,7 +278,7 @@ public class CoreSickle extends ToolItem implements ICoreTool {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override