+++ src/main/java/net/silentchaos512/gear/item/gear/GearFishingRod.java
@@ -136,7 +136,7 @@ public class GearFishingRod extends FishingRodItem implements ICoreTool {
 
     @Override
     public int getItemEnchantability(ItemStack stack) {
-        return GearData.getStatInt(stack, ItemStats.ENCHANTABILITY);
+        return GearHelper.getEnchantability(stack);
     }
 
     @Override