+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Two config options related to gear enchanting (both in `gear.enchanting`). One allows or disallows normal enchanting (sets enchantability to zero). The other will forcibly remove all enchantments from existing items except those added by traits. [#375]
 ### Changed
 - Copper's stats
 ### Fixed