+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.30] - 2021-07-12
+This version _does NOT_ load with Silent Lib 4.10.x! Continue using 4.9.6 for now.
 ### Added
 - Deepslate bort ore (unused, 1.17 prep)
+### Removed
+- Silent Gear's old `ExclusionIngredient` type (replaced by Silent Lib equivalent)
 ### Changed
 - Bort ore textures tweaked